+++ src/main/java/com/account/yomankum/accountbook/service/AccountBookService.java
@@ -1,0 +1,9 @@
+package com.account.yomankum.accountbook.service;
+
+import com.account.yomankum.accountbook.domain.AccountWriteDto;
+
+import java.util.List;
+
+public interface AccountBookService {
+    void write(List<AccountWriteDto> accountWriteDtoList);
+}
