+++ src/main/java/com/account/yomankum/domain/SnsUser.java
@@ -1,6 +1,6 @@
 package com.account.yomankum.domain;
 
-import com.account.yomankum.login.security.oauth.Sns;
+import com.account.yomankum.security.oauth.Sns;
 import jakarta.persistence.*;
 import lombok.*;
