--- src/main/java/com/account/yomankum/login/security/oauth/JwtValue.java
@@ -1,14 +1,0 @@
-package com.account.yomankum.login.security.oauth;
-
-public interface JwtValue {
-
-    String getFirstKid();
-    String getSecondKid();
-    String getKid();
-    String getKty();
-    String getAlg();
-    String getUse();
-    String getN();
-    String getE();
-    void jwkSetting(String number);
-}
