--- src/main/java/com/account/yomankum/login/security/oauth/NaverUserInfo.java
@@ -1,29 +1,0 @@
-package com.account.yomankum.login.security.oauth;
-
-import java.util.Map;
-
-public class NaverUserInfo implements SnsUserInfo{
-
-    private final Map<String, Object> attributes;
-
-    public NaverUserInfo(Map<String, Object> attributes) {
-        this.attributes = attributes;
-    }
-
-    @Override
-    public String getUUIDKey() {
-        return String.valueOf(attributes.get("id"));
-    }
-
-    @Override
-    public Sns getSnsName() {
-        return Sns.NAVER;
-    }
-
-    @Override
-    public String getEmail() {
-        return String.valueOf(attributes.get("email"));
-    }
-
-}
-
