--- src/main/java/com/account/yomankum/login/security/oauth/Sns.java
@@ -1,6 +1,0 @@
-package com.account.yomankum.login.security.oauth;
-
-public enum Sns {
-
-    NAVER, KAKAO, GOOGLE
-}
