--- src/main/java/com/account/yomankum/login/security/oauth/SnsUserInfo.java
@@ -1,9 +1,0 @@
-package com.account.yomankum.login.security.oauth;
-
-public interface SnsUserInfo {
-
-    String getUUIDKey();
-    Sns getSnsName();
-    String getEmail();
-
-}
