+++ src/main/java/com/account/yomankum/repository/SnsUserRepository.java
@@ -1,6 +1,6 @@
 package com.account.yomankum.repository;
 
-import com.account.yomankum.login.security.oauth.Sns;
+import com.account.yomankum.security.oauth.Sns;
 import com.account.yomankum.domain.SnsUser;
 import org.springframework.data.jpa.repository.JpaRepository;
