+++ src/main/java/com/account/yomankum/security/jwt/TokenService.java
@@ -1,0 +1,14 @@
+package com.account.yomankum.security.jwt;
+
+import com.account.yomankum.domain.Name;
+
+public interface TokenService {
+
+    String creatToken(Long id, String nickname, Name name);
+    String createRefreshToken();
+    String reCreateToken(String token);
+    boolean tokenValid(String token);
+    String getNicknameByToken(String token);
+    Long getIdByToken(String token);
+    String getSnsUUID(String sns, String token);
+}
