+++ src/main/java/com/account/yomankum/security/oauth/JwtValue.java
@@ -1,0 +1,14 @@
+package com.account.yomankum.security.oauth;
+
+public interface JwtValue {
+
+    String getFirstKid();
+    String getSecondKid();
+    String getKid();
+    String getKty();
+    String getAlg();
+    String getUse();
+    String getN();
+    String getE();
+    void jwkSetting(String number);
+}
