+++ src/main/java/com/account/yomankum/security/oauth/Sns.java
@@ -1,0 +1,6 @@
+package com.account.yomankum.security.oauth;
+
+public enum Sns {
+
+    NAVER, KAKAO, GOOGLE
+}
