+++ src/main/java/com/account/yomankum/security/oauth/SnsUserInfo.java
@@ -1,0 +1,9 @@
+package com.account.yomankum.security.oauth;
+
+public interface SnsUserInfo {
+
+    String getUUIDKey();
+    Sns getSnsName();
+    String getEmail();
+
+}
