+++ src/main/java/com/account/yomankum/security/oauth/TokenResponse.java
@@ -1,0 +1,32 @@
+package com.account.yomankum.security.oauth;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+public class TokenResponse {
+
+    @JsonProperty("token_type")
+    private String tokenType;
+    @JsonProperty("access_token")
+    private String accessToken;
+    @JsonProperty("id_token")
+    private String idToken;
+    @JsonProperty("expires_in")
+    private long expiresIn;
+    @JsonProperty("refresh_token")
+    private String refreshToken;
+    @JsonProperty("refresh_token_expires_in")
+    private long refreshTokenExpiresIn;
+    @JsonProperty("error")
+    private String error;
+    @JsonProperty("error_description")
+    private String errorDescription;
+    @JsonProperty("scope")
+    private String scope; // 여러개 올 경우, 공백으로 구분 "aaa bbb ccc"
+
+
+}
+
