--- src/main/java/com/account/yomankum/service/AccountBookService.java
@@ -1,9 +1,0 @@
-package com.account.yomankum.service;
-
-import com.account.yomankum.domain.dto.AccountWriteDto;
-
-import java.util.List;
-
-public interface AccountBookService {
-    void write(List<AccountWriteDto> accountWriteDtoList);
-}
